refactor(slack_app): move the fork activity into the activities package
Every other Slack-app activity lives in `activities/` and is registered through
`SLACK_APP_ACTIVITIES`; the fork one sat beside its workflow and was registered
separately. It now follows the same shape, leaving `slack_app_fork.py` holding
only the workflow.

`services/fork_context.py` becomes `slack_fork_context.py`, matching the
`slack_`-prefixed modules around it.
