+++ posthog/temporal/ai/__init__.py
@@ -11,10 +11,7 @@ from posthog.temporal.ai.slack_app import SLACK_APP_ACTIVITIES
 from posthog.temporal.ai.slack_app.posthog_code_slack_mention import PostHogCodeSlackMentionWorkflow
 from posthog.temporal.ai.slack_app.posthog_code_slack_mention_command import PostHogCodeSlackMentionCommandWorkflow
 from posthog.temporal.ai.slack_app.posthog_slack_inbox_onboarding import PostHogSlackInboxOnboardingWorkflow
-from posthog.temporal.ai.slack_app.slack_app_fork import (
-    SlackAppForkThreadWorkflow,
-    process_slack_app_fork_thread_activity,
-)
+from posthog.temporal.ai.slack_app.slack_app_fork import SlackAppForkThreadWorkflow
 from posthog.temporal.ai.slack_app.slack_app_mention import SlackAppMentionWorkflow
 
 from .llm_traces_summaries.summarize_traces import (
