+++ posthog/temporal/ai/slack_app/__init__.py
@@ -30,6 +30,7 @@ from posthog.temporal.ai.slack_app.activities import (
     post_posthog_code_no_repos_activity,
     post_posthog_code_picker_timeout_activity,
     post_posthog_code_repo_picker_activity,
+    process_slack_app_fork_thread_activity,
     request_untagged_followup_confirmation_activity,
     resolve_posthog_code_slack_command_user_activity,
     run_posthog_slack_inbox_onboarding_activity,
@@ -51,6 +52,7 @@ from posthog.temporal.ai.slack_app.types import (
 )
 
 SLACK_APP_ACTIVITIES = [
+    process_slack_app_fork_thread_activity,
     classify_untagged_followup_activity,
     request_untagged_followup_confirmation_activity,
     resolve_posthog_code_slack_command_user_activity,
@@ -89,6 +91,7 @@ __all__ = [
     "PostHogCodeSlackMentionWorkflowInputs",
     "PostHogSlackInboxOnboardingInputs",
     "SLACK_APP_ACTIVITIES",
+    "process_slack_app_fork_thread_activity",
     "SlackAppMentionWorkflowInputs",
     "SlackAppModelOverride",
     "SlackAppModelOverrideInput",
